Bastion must use Flatcar Linux
CoreOS has been EOL since early 2020, so we plan to use Flatcar Linux, which
is a direct fork of CoreOS.
Also, we must revert Lance's two changes to how SSH fingerprints are stored
and pulled, which was causing issues authenticating users.
These changes were tested and found to be effective in restoring bastion
access.

Ticker: SEC-418